Fix some problems in method_handler
The custom sanitize function is replaced with g_strescape, which
allocates the target buffer. This makes buffer overflows impossible and
in this case, replaces the shared buffer sanitized. This is needed,
because it is used on two occasions for two values at the same time,
thereby overwriting the first value (app_name/app_icon and
summary/body). This bug is also mention in issue #10.

Additionally, several memleaks in this function are fixed (free and
unref results of the GVariant functions).