Always add a newline after "hints:" and "actions:" when not using PRINT_JSON mode
This way, the next line is not fused with the hints and actions line,
when the is no hint or action.

Before this commit with action "foo: bar" and summary "baz":
```
actions:	foo: baz
summary: baz
```

After this commit:
```
actions:
	foo: baz
summary: baz
```

Before this commit without action and summary "baz":
```
actions: summary: baz
```

After this commit:
```
actions:
summary: baz
```

Fixes #14.