Add a formatter for values of unknown type
The hints_output_iterator() function skips a hint if its value is not of
any of the types explicitly checked for in the if-else chain. However,
it prints a comma before this check, meaning values of unknown type
result in a trailing comma. By adding a formatter for such values, we
not only resolve the trailing comma issue, but also let the user know
that their notification contained a hint of an unknown type, rather than
hiding it.